test(pipnn): colocate HashPrune coverage
Adapt HashPrune to current main, fix its concurrent aliasing boundaries, and keep all private/composition tests beside their implementation.
